валидация для соотношения цены и типа жилья,времени въезда и выезда,длины строки,соотношения количества комнат и гостей
